Change name principal
