+++ Modern-UI-Design-C-Sharp/FormDesign/FormMain.cs
@@ -1,0 +1,77 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Linq;
+using System.Runtime.InteropServices;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace Modern_UI_Design_C_Sharp.FormDesign
+{
+    public partial class FormMain : Form
+    {
+        //Fields
+        private int borderSize = 2;
+        private Size formSize; //Keep form size when it is minimized and restored.Since the form is resized because it takes into account the size of the title bar and borders.
+
+        [DllImport("user32.dll", EntryPoint = "ReleaseCapture")]
+        private extern static void ReleaseCapture();
+
+        [DllImport("user32.DLL", EntryPoint = "SendMessage")]
+        private extern static void SendMessage(System.IntPtr hWnd, int wMsg, int wParam, int lParam);
+        
+        public FormMain()
+        {
+            InitializeComponent();
+            this.Padding = new Padding(borderSize);//Border size
+            this.BackColor = Color.FromArgb(98, 102, 244);//Border color
+        }
+
+        private void iconButtonClose_Click(object sender, EventArgs e)
+        {
+            Application.Exit();
+        }
+
+        private void iconButtonNormal_Click(object sender, EventArgs e)
+        {
+            iconButtonNormal.Visible = false;
+            iconButtonPlus.Visible = true;
+            this.WindowState = FormWindowState.Normal;
+        }
+
+        private void iconButtonMinus_Click(object sender, EventArgs e)
+        {
+            this.WindowState = FormWindowState.Minimized;
+        }
+
+        private void iconButtonPlus_Click(object sender, EventArgs e)
+        {
+            this.WindowState = FormWindowState.Maximized;
+            iconButtonNormal.Visible = true;
+            iconButtonPlus.Visible = false;
+        }
+
+        private void PanelFormBar_MouseDown(object sender, MouseEventArgs e)
+        {
+            ReleaseCapture();
+            SendMessage(this.Handle, 0x112, 0xf012, 0);
+        }
+
+        //Overridden methods
+        protected override void WndProc(ref Message m)
+        {
+            const int WM_NCCALCSIZE = 0x0083;//Standar Title Bar - Snap Window
+           
+
+            //Remove border and keep snap window
+            if (m.Msg == WM_NCCALCSIZE && m.WParam.ToInt32() == 1)
+            {
+                return;
+            }
+            base.WndProc(ref m);
+        }
+    }
+}
