--- Modern-UI-Design-C-Sharp/FormDesign/FormMaster.cs
@@ -1,73 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Runtime.InteropServices;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace Modern_UI_Design_C_Sharp
-{
-    public partial class FormMaster : Form
-    {
-        
-        [DllImport("user32.dll", EntryPoint = "ReleaseCapture")]
-        private extern static void ReleaseCapture();
-
-        [DllImport("user32.DLL", EntryPoint = "SendMessage")]
-        private extern static void SendMessage(System.IntPtr hWnd, int wMsg, int wParam, int lParam);
-        private void header_MouseDown(object sender, MouseEventArgs e)
-        {
-            ReleaseCapture();
-            SendMessage(this.Handle, 0x112, 0xf012, 0);
-        }
-        [DllImport("Gdi32.dll", EntryPoint = "CreateRoundRectRgn")]
-        private static extern IntPtr CreateRoundRectRgn
-        (
-            int nLeftRect,     // x-coordinate of upper-left corner
-            int nTopRect,      // y-coordinate of upper-left corner
-            int nRightRect,    // x-coordinate of lower-right corner
-            int nBottomRect,   // y-coordinate of lower-right corner
-            int nWidthEllipse, // width of ellipse
-            int nHeightEllipse // height of ellipse
-        );
-        public FormMaster()
-        {
-            InitializeComponent();
-            this.SetStyle(ControlStyles.ResizeRedraw, true);
-            this.DoubleBuffered = true;
-            this.FormBorderStyle = FormBorderStyle.None;
-            Region = System.Drawing.Region.FromHrgn(CreateRoundRectRgn(0, 0, Width, Height, 10, 10));
-        }
-        
-        private void iconButtonClose_Click(object sender, EventArgs e)
-        {
-            this.Close();
-        }
-
-        private void iconButtonMinus_Click(object sender, EventArgs e)
-        {
-            this.WindowState = FormWindowState.Minimized;
-        }
-
-        private void iconButtonPlus_Click(object sender, EventArgs e)
-        {
-            Region =  new Region();
-            this.WindowState = FormWindowState.Maximized;
-            iconButtonNormal.Visible = true;
-            iconButtonPlus.Visible = false;
-        }
-
-        private void iconButtonNormal_Click(object sender, EventArgs e)
-        {
-            iconButtonNormal.Visible = false;
-            iconButtonPlus.Visible = true;
-            this.WindowState = FormWindowState.Normal;            
-            Region = System.Drawing.Region.FromHrgn(CreateRoundRectRgn(0, 0, Width, Height, 10, 10));
-        }
-
-    }
-}
