+++ Modern-UI-Design-C-Sharp/Modern-UI-Design-C-Sharp.csproj
@@ -52,16 +52,16 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Form\FormMaster.cs">
+    <Compile Include="FormDesign\FormMain.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Form\FormMaster.Designer.cs">
-      <DependentUpon>FormMaster.cs</DependentUpon>
+    <Compile Include="FormDesign\FormMain.Designer.cs">
+      <DependentUpon>FormMain.cs</DependentUpon>
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <EmbeddedResource Include="Form\FormMaster.resx">
-      <DependentUpon>FormMaster.cs</DependentUpon>
+    <EmbeddedResource Include="FormDesign\FormMain.resx">
+      <DependentUpon>FormMain.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
